Add files via upload
RPLIDAR_directly_Controlled_by_ESP32.ino and  my_F_matrix.m are use together to test the kinematic linear transformations 

here is the procedure I used:
with the RPLIDAR connected to a computer and using the arduino serial port
1. I type  t <enter> in the serial monitor windowf
2. I type x <enter> in the serial monitor window
3. I also type 0 <enter> in the serial monitor window which clears the measurement array- it needs to be cleared before typing the the t command again

I ran this procedure three times with the RPLIDAR in three different locations once at the origin, one move forward 20 cm and one moved forward 20 cm and then rotated 90 degree clockwise from the top view.

the translation and rotation information is hard coded into the m file. 

the outputs absolute_point_xx (1a through 4c) take the meaurements taken from the arduino (which are relative to it's current location, and converts them to absolute coordinates.

further testing is needed, but this is a good starting point and I wanted to save it as a reference on how the Math works for when I write the code in C or python later.